Update axiodb dependency to version 1.7.10, remove Fastify server implementation, and restructure configuration keys for better organization.
