+++ src/config/Keys.ts
@@ -4,3 +4,21 @@ export enum Port {
     GRPC = 27020,
     WEBSOCKET = 27021
 }
+
+export enum FastifyInformation {
+    CentralDB_Name = "Fastify_AxioDB_Information",
+    Custompath= "./Central"
+}
+
+export enum TCPInformation {
+    CentralDB_Name = "TCP_AxioDB_Information",
+    Custompath= "./Central"
+}
+export enum GRPCInformation {
+    CentralDB_Name = "GRPC_AxioDB_Information",
+    Custompath= "./Central"
+}
+export enum WebSocketInformation {
+    CentralDB_Name = "WebSocket_AxioDB_Information",
+    Custompath= "./Central"
+}
